Attach Modal discovery credentials
