+++ .github/scripts/sync_modal_secrets.py
@@ -42,6 +42,20 @@ def main() -> None:
             "--env",
             env_name,
             "--force",
+            "modal-token",
+            f"MODAL_TOKEN_ID={os.environ['MODAL_TOKEN_ID']}",
+            f"MODAL_TOKEN_SECRET={os.environ['MODAL_TOKEN_SECRET']}",
+        ],
+        check=True,
+    )
+    subprocess.run(
+        [
+            "modal",
+            "secret",
+            "create",
+            "--env",
+            env_name,
+            "--force",
             "gcp-credentials",
             (
                 "GOOGLE_APPLICATION_CREDENTIALS_JSON="
