+++ .github/workflows/pipeline.yaml
@@ -132,6 +132,7 @@ jobs:
           PARALLEL_MATRIX: ${{ inputs.parallel_matrix || 'false' }}
           NUM_MATRIX_WORKERS: ${{ inputs.num_matrix_workers || '50' }}
         run: |
+          modal secret create --env="${MODAL_ENVIRONMENT}" --force modal-token MODAL_TOKEN_ID="${MODAL_TOKEN_ID}" MODAL_TOKEN_SECRET="${MODAL_TOKEN_SECRET}"
           modal deploy --env="${MODAL_ENVIRONMENT}" modal_app/pipeline_discovery.py
           modal deploy --env="${MODAL_ENVIRONMENT}" --name="${US_DATA_MODAL_APP_NAME}" --tag="${US_DATA_RUN_ID}" modal_app/pipeline.py
           python .github/scripts/spawn_modal_pipeline.py
