+++ modal_app/pipeline.py
@@ -545,6 +545,7 @@ def verify_runtime_seams() -> dict:
         "modal_app/local_area.py",
         "modal_app/pipeline_discovery.py",
         "modal_app/pipeline_discovery_core.py",
+        "modal_app/pipeline_discovery_schema.py",
         "modal_app/pipeline_status.py",
         "modal_app/h5_test_harness.py",
         "modal_app/step_manifests/specs.py",
@@ -596,6 +597,7 @@ def verify_runtime_seams() -> dict:
         "modal_app.local_area",
         "modal_app.pipeline_discovery",
         "modal_app.pipeline_discovery_core",
+        "modal_app.pipeline_discovery_schema",
         "modal_app.pipeline_status",
         "modal_app.remote_calibration_runner",
         "modal_app.step_manifests.specs",
